backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Task Scheduling
apscheduler==3.10.4

# AI and Agent Framework
langchain==0.1.5
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# HTTP Requests
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
pydantic==2.5.3
python-multipart==0.0.6
